% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matTraining}
\alias{matTraining}
\title{Training Expression Data}
\format{
A numeric matrix with 70 rows (genes) and 78 columns (samples).
}
\source{
Glas et al., BMC Genomics, 2006. Originally obtained from the
MammaPrintData Bioconductor package as described in Marchionni et al.,
BMC Genomics, 2013.
}
\usage{
matTraining
}
\description{
A numeric matrix of gene expression data for 78 breast cancer patients and
70 genes from the MammaPrint assay. Samples are in columns and genes in
rows.
}
\examples{
data(trainingData)
dim(matTraining)
table(trainingGroup)
}
\keyword{datasets}
