% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{swap_plot_votes}
\alias{swap_plot_votes}
\title{Heatmap of TSP Votes}
\usage{
swap_plot_votes(classifier, x, y = NULL, combine_fn)
}
\arguments{
\item{classifier}{A KTSP classifier.}

\item{x}{Numeric matrix (features x samples).}

\item{y}{Optional factor of class labels (used for column annotation).}

\item{combine_fn}{Optional combine function passed to
\code{\link[=swap_ktsp_statistics]{swap_ktsp_statistics()}}.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Creates a heatmap of the individual TSP comparison votes per sample.
}
\examples{
data(trainingData)
classifier <- swap_train_ktsp(matTraining, trainingGroup, k_range = 3:8)
swap_plot_votes(classifier, matTraining, trainingGroup)
}
