% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{swap_score_vector_to_matrix}
\alias{swap_score_vector_to_matrix}
\title{Convert Named Score Vector to Matrix}
\usage{
swap_score_vector_to_matrix(v)
}
\arguments{
\item{v}{A named numeric vector.}
}
\value{
A numeric matrix with appropriate rownames and colnames.
}
\description{
Converts a named vector of scores (with \code{"gene1,gene2"} names) back to
a matrix.
}
\examples{
v <- c("a,b" = 1, "a,c" = 2, "b,a" = 3, "b,c" = 4)
swap_score_vector_to_matrix(v)
}
