% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testingGroup}
\alias{testingGroup}
\title{Test Class Labels}
\format{
A factor of length 307 with levels \code{"bad"} and \code{"good"}.
}
\source{
Buyse et al., JNCI, 2006.
}
\usage{
testingGroup
}
\description{
A factor with two levels (\code{"bad"} and \code{"good"}) indicating the prognosis
group for each of the 307 test samples.
}
\examples{
data(testingData)
table(testingGroup)
}
\keyword{datasets}
