% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trainingGroup}
\alias{trainingGroup}
\title{Training Class Labels}
\format{
A factor of length 78 with levels \code{"bad"} and \code{"good"}.
}
\source{
Glas et al., BMC Genomics, 2006.
}
\usage{
trainingGroup
}
\description{
A factor with two levels (\code{"bad"} and \code{"good"}) indicating the prognosis
group for each of the 78 training samples.
}
\examples{
data(trainingData)
table(trainingGroup)
}
\keyword{datasets}
