% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchbox-package.R
\docType{package}
\name{switchbox-package}
\alias{switchbox}
\alias{switchbox-package}
\title{switchbox: K-Top-Scoring-Pair (KTSP) Classification}
\description{
Train and validate binary classifiers based on pair-wise feature comparisons using the K-Top-Scoring-Pair (KTSP) algorithm. The KTSP classifier uses the relative ordering of feature values (e.g. gene expression levels) to make predictions, making it robust to data normalization and producing transparent decision rules. The package supports feature filtering, restricted pair selection, cross-validation, and various performance evaluation methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/luciorq/switchbox}
  \item Report bugs at \url{https://github.com/luciorq/switchbox/issues}
}

}
\author{
\strong{Maintainer}: Lucio Rezende Queiroz \email{luciorqueiroz@gmail.com}

Authors:
\itemize{
  \item Bahman Afsari \email{bahman@jhu.edu} [copyright holder]
  \item Luigi Marchionni \email{marchion@jhu.edu} [copyright holder]
  \item Wikum Dinalankara \email{wdinala1@jhmi.edu} [copyright holder]
}

}
\keyword{internal}
